Add variable for filtered notes list.
Eventually this functionality will be the same as the filtered tag list.
When the user searches, the search-field widget will do the work of
grepping for the files that match and setting this variable. (This is
stubbed out for now with a quick and dirty version.)

Then, when calling the list-files function, the menu-setup function pass
the list of files to be displayed. Instead of having the file-finding
and file-listing process bound up together, the list-files function will
become single-purpose, of just displaying the widgets for the files it's
passed.